@@ -27,49 +27,47 @@
     "author": "collaboration Factory AG",
     "license": "ISC",
     "dependencies": {
-        "@openapitools/openapi-generator-cli": "2.1.26",
-        "chokidar": "^3.0.2",
-        "copyfiles": "^2.1.0",
-        "cross-spawn": "^6.0.5",
-        "css-loader": "^5.1.3",
-        "css-minimizer-webpack-plugin": "^1.3.0",
+        "@openapitools/openapi-generator-cli": "^2.5.2",
+        "chokidar": "^3.5.3",
+        "copyfiles": "^2.4.1",
+        "cross-spawn": "^7.0.3",
+        "css-loader": "^6.6.0",
+        "css-minimizer-webpack-plugin": "^4.2.2",
         "eol-converter-cli": "^1.0.8",
         "expose-loader": "^1.0.3",
         "file-loader": "^6.2.0",
-        "generic-pool": "latest",
+        "generic-pool": "3.9.0",
         "imports-loader": "^1.2.0",
         "less": "^3.8.1",
         "less-loader": "^7.3.0",
-        "meow": "^5.0.0",
-        "mini-css-extract-plugin": "^1.3.9",
-        "raw-loader": "^1.0.0",
-        "rimraf": "^2.6.2",
-        "sass": "^1.43.4",
-        "sass-loader": "^10.1.1",
+        "meow": "^9.0.0",
+        "mini-css-extract-plugin": "^1.6.2",
+        "raw-loader": "^4.0.2",
+        "rimraf": "^3.0.2",
+        "sass": "^1.49.7",
+        "sass-loader": "^12.4.0",
         "script-loader": "^0.7.2",
-        "source-map-loader": "^0.2.4",
+        "source-map-loader": "^3.0.1",
+        "terser-webpack-plugin": "^5.3.6",
         "uglify-loader": "^3.0.0",
-        "uglifyjs-webpack-plugin": "^2.2.0",
-        "webpack": "^4.37.0",
+        "webpack": "^5.68.0",
         "webpack-merge": "^5.8.0",
-        "xml2js": "^0.4.19"
+        "xml2js": "^0.4.23"
     },
     "devDependencies": {
-        "@types/copyfiles": "^2.1.0",
-        "@types/cross-spawn": "^6.0.0",
-        "@types/generic-pool": "^3.1.6",
+        "@types/copyfiles": "^2.4.1",
+        "@types/cross-spawn": "^6.0.2",
+        "@types/generic-pool": "^3.1.10",
         "@types/jest": "^27.0.3",
-        "@types/less": "^3.0.0",
-        "@types/meow": "^5.0.0",
-        "@types/minimist": "^1.2.0",
-        "@types/node": "^14.18.31",
-        "@types/rimraf": "^2.0.2",
-        "@types/vinyl": "^2.0.2",
-        "@types/webpack": "^4.4.19",
-        "jest": "^27.4.5",
+        "@types/less": "^3.0.3",
+        "@types/minimist": "^1.2.2",
+        "@types/node": "^18.11.0",
+        "@types/rimraf": "^3.0.2",
+        "@types/vinyl": "^2.0.6",
+        "jest": "^27.5.1",
         "prettier": "2.5.1",
         "tmp": "^0.2.1",
-        "ts-jest": "^27.1.2",
-        "typescript": "^3.9.10"
+        "ts-jest": "^27.1.3",
+        "typescript": "^4.8.4"
     }
 }
