@@ -4,7 +4,7 @@ import { lessEscapePlugin } from './LessEscapePlugin';
  * Exports all the less plugins
  * @param pluginName Plugin name
  */
-export function lessPlugins(pluginName: string) {
+export function lessPlugins(pluginName: string): Less.Plugin[] {
     return [
         {
             install: (less, pluginManager) => {
